@@ -1,5 +1,5 @@
 ﻿using System.Linq;
-using Lykke.NuGetReferencesScanner.Domain;
+using Lykke.NuGetReferencesScanner.Domain.Abstractions;
 using Lykke.NuGetReferencesScanner.Models;
 using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
