@@ -1,8 +1,9 @@
 ﻿using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using Lykke.NuGetReferencesScanner.Domain.Models;
 
-namespace Lykke.NuGetReferencesScanner.Domain
+namespace Lykke.NuGetReferencesScanner.Domain.Abstractions
 {
     internal interface IOrganizationScanner
     {
