@@ -2,6 +2,8 @@
 using System.IO;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Hosting;
 
 namespace Lykke.NuGetReferencesScanner
 {
@@ -21,12 +23,23 @@ namespace Lykke.NuGetReferencesScanner
             }
         }
 
-        private static IWebHost BuildWebHost(string[] args) =>
-            new WebHostBuilder()
-            .UseKestrel()
-            .UseUrls("http://*:5000")
-            .UseContentRoot(Directory.GetCurrentDirectory())
-            .UseStartup<Startup>()
-            .Build();
+        private static IHost BuildWebHost(string[] args) =>
+            Host.CreateDefaultBuilder(args)
+                .ConfigureAppConfiguration(builder =>
+                {
+                    builder
+                        .AddJsonFile("appsettings.json", optional: true, reloadOnChange: true)
+                        .AddEnvironmentVariables();
+                })
+                .ConfigureWebHostDefaults(webBuilder =>
+                {
+                    webBuilder.UseStartup<Startup>();
+                    webBuilder.ConfigureKestrel((ctx, opts) =>
+                    {
+                        webBuilder.UseUrls("http://*:5000");
+                    });
+                    webBuilder.UseContentRoot(Directory.GetCurrentDirectory());
+                })
+                .Build();
     }
 }
